SCRD-2497 restart verb and switch to state: present
For maintenance update support, we need a verb
that can be used to enforce an unconditional restart
of all services, to be used either to orchestrate
updates manually or to trigger a cloud-wide restart
of services after an update via the ardana-update.yml
or the ardana-restart.yml playbooks.

This patch also replaces state: latest with state: present,
because we don't want packages to be updated by the install
playbooks, we want to orchestrate package updates globally.

Change-Id: I5edbb682abf0ee8e92f5809a48368f77d1434262